+++ packages/block-editor/src/components/colors-gradients/panel-color-gradient-settings.js
@@ -42,6 +42,7 @@ export const PanelColorGradientSettingsInner = ( {
 	const panelId = useInstanceId( PanelColorGradientSettingsInner );
 	const { batch } = useRegistry();
 	if (
+		! children &&
 		( ! colors || colors.length === 0 ) &&
 		( ! gradients || gradients.length === 0 ) &&
 		disableCustomColors &&