+++ packages/block-editor/src/components/panel-color-settings/test/index.js
@@ -1,7 +1,7 @@
 /**
  * External dependencies
  */
-import { render } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 
 /**
  * Internal dependencies
@@ -32,6 +32,20 @@ describe( 'PanelColorSettings', () => {
 			/>
 		);
 		expect( container ).toBeEmptyDOMElement();
+	} );
+
+	it( 'should render children even if there are no colors to choose', async () => {
+		render(
+			<PanelColorSettings
+				title="Test Title"
+				colors={ [] }
+				disableCustomColors
+				colorSettings={ [] }
+			>
+				<p>Additional setting</p>
+			</PanelColorSettings>
+		);
+		expect( screen.getByText( 'Additional setting' ) ).toBeInTheDocument();
 	} );
 
 	it( 'should render a color panel if at least one setting supports custom colors', async () => {